Draft of POC for integration testing
