+++ rsky-pds/Cargo.toml
@@ -11,44 +11,44 @@ repository = "https://github.com/blacksky-algorithms/rsky/tree/main/rsky-pds"
 documentation = "https://docs.rs/rsky-pds"
 
 [dependencies]
-tokio = {workspace = true}
-rocket = { version = "=0.5.1", features = ["json","tls"] }
+tokio = { workspace = true }
+rocket = { version = "=0.5.1", features = ["json", "tls"] }
 dotenvy = "0.15"
 rsky-lexicon = { workspace = true }
 rsky-identity = { workspace = true }
 rsky-crypto = { workspace = true }
-rsky-common = {workspace = true }
+rsky-common = { workspace = true }
 rsky-syntax = { workspace = true }
 rsky-repo = { workspace = true }
 diesel = { version = "=2.1.5", features = ["chrono", "postgres"] }
 chrono = "0.4.26"
 serde = { workspace = true }
 serde_repr = "0.1"
-serde_derive = {workspace = true}
-rand = {workspace = true}
+serde_derive = { workspace = true }
+rand = { workspace = true }
 email_address = "0.2.4"
 anyhow = "1.0.79"
 serde_cbor = { workspace = true }
 base64 = "0.22.0"
 data-encoding = "2.5.0"
-reqwest = { version = "0.12.3",features = ["json","blocking"] }
-serde_json = {workspace = true}
+reqwest = { version = "0.12.3", features = ["json", "blocking"] }
+serde_json = { workspace = true }
 serde_ipld_dagcbor = { workspace = true }
 serde_bytes = { workspace = true }
 base64-url = "2.0.2"
-secp256k1 = {workspace = true}
-rand_core = {workspace = true}
-sha2 = {workspace = true}
-indexmap = { version = "1.9.3",features = ["serde-1"] }
+secp256k1 = { workspace = true }
+rand_core = { workspace = true }
+sha2 = { workspace = true }
+indexmap = { version = "1.9.3", features = ["serde-1"] }
 hex = "0.4.3"
 libipld = { workspace = true }
 lazy_static = "1.4.0"
 regex = "1.10.3"
 thiserror = "1.0.40"
 aws-sdk-s3 = "1.29.0"
-aws-config = { version = "1.1.8",features = ["behavior-version-latest"] }
+aws-config = { version = "1.1.8", features = ["behavior-version-latest"] }
 futures = "0.3.28"
-jwt-simple = { version = "0.12.9",default-features = false, features = ["pure-rust"] }
+jwt-simple = { version = "0.12.9", default-features = false, features = ["pure-rust"] }
 argon2 = "0.5.3"
 base64ct = "1.6.0"
 mailgun-rs = "0.1.10"
@@ -59,7 +59,7 @@ toml = "0.8.12"
 ws = { package = "rocket_ws", version = "0.1.1" }
 atrium-api = "0.24.6"
 atrium-xrpc-client = "0.5.8"
-atrium-ipld = {package = "ipld-core", version = "0.4.1"}
+atrium-ipld = { package = "ipld-core", version = "0.4.1" }
 time = "^0.3.36"
 url = "2.5.2"
 async-event-emitter = "0.1.3"
@@ -71,7 +71,11 @@ tracing-subscriber = "0.3.19"
 [dev-dependencies]
 testcontainers = "0.23.2"
 testcontainers-modules = { version = "0.11.6", features = ["postgres", "blocking"] }
+diesel_migrations = {version = "2.1.0", features = ["postgres"]}
+http-auth-basic = { version = "0.3.5" }
 
 [dependencies.rocket_sync_db_pools]
 version = "=0.1.0"
 features = ["diesel_postgres_pool"]
+
+
